feat: finish payment result page and edit navigate path in login form and register page
